Move annotation property tool out of the annotation/index.ts. Moved serialized tool identifiers into tool_ids.ts.
Add tests covering enum property tool bindings.
